Changed coordinate type from `double` to `float`

+++ raytracing/src/raytracing/Body.java
@@ -6,10 +6,10 @@ package raytracing;
  */
 public interface Body {
     // Implicit function that defines the body
-    double f(double x, double y, double z);
+    float f(float x, float y, float z);
     
     // Partial derivations of f
-    double fx(double x, double y, double z);
-    double fy(double x, double y, double z);
-    double fz(double x, double y, double z);
+    float fx(float x, float y, float z);
+    float fy(float x, float y, float z);
+    float fz(float x, float y, float z);
 }
